Reorganize providers
