+++ extension/ui/pages/routes.tsx
@@ -1,6 +1,3 @@
-import { SidebarProvider } from "@app/components/sparkle/SidebarContext";
-import { LinkWrapper } from "@app/lib/platform";
-import { SparkleContext } from "@dust-tt/sparkle";
 import { ProtectedRoute } from "@extension/ui/components/auth/ProtectedRoute";
 import { LoginPage } from "@extension/ui/pages/LoginPage";
 import { MainPage } from "@extension/ui/pages/MainPage";
@@ -37,17 +34,11 @@ export const routes = [
     element: (
       <ProtectedRoute>
         {({ user, workspace, handleLogout }) => (
-          <SparkleContext.Provider
-            value={{ components: { link: LinkWrapper } }}
-          >
-            <SidebarProvider>
-              <MainPage
-                user={user}
-                workspace={workspace}
-                handleLogout={handleLogout}
-              />
-            </SidebarProvider>
-          </SparkleContext.Provider>
+          <MainPage
+            user={user}
+            workspace={workspace}
+            handleLogout={handleLogout}
+          />
         )}
       </ProtectedRoute>
     ),
